requirements.txt
asgiref==3.8.1
camelcase==0.2
click==8.1.7
colorama==0.4.6
Django==5.1.5
django-browser-reload==1.17.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.6
psycopg2==2.9.10
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.3
tzdata==2024.2
whitenoise==6.8.2
